Updates for paper

+++ aligner/archive.py
@@ -1,5 +1,3 @@
-
-
 import os
 import pickle
 
@@ -11,7 +9,6 @@ FORMAT = "zip"
 
 
 class Archive(object):
-
     """
     Class representing data in a directory or archive file (zip, tar,
     tar.gz/tgz)
@@ -42,7 +39,7 @@ class Archive(object):
         """
         base = mkdtemp(dir=os.environ.get("TMPDIR", None))
         source = os.path.join(base, head)
-        os.makedirs(source, exist_ok = True)
+        os.makedirs(source, exist_ok=True)
         return cls(source, True)
 
     def add(self, source):
@@ -71,22 +68,22 @@ class Archive(object):
         """
         Add file into archive
         """
-        os.makedirs(destination, exist_ok = True)
+        os.makedirs(destination, exist_ok=True)
         ali_model_path = os.path.join(self.dirname, 'ali-final.mdl')
-        if os.path.exists(ali_model_path):
+        if False and os.path.exists(ali_model_path):
             copyfile(ali_model_path, os.path.join(destination, 'final.mdl'))
             copyfile(os.path.join(self.dirname, 'ali-final.occs'), os.path.join(destination, 'final.occs'))
             copyfile(os.path.join(self.dirname, 'ali-tree'), os.path.join(destination, 'tree'))
         else:
-            copyfile(os.path.join(self.dirname, 'final.mdl'),  os.path.join(destination, 'final.mdl'))
+            copyfile(os.path.join(self.dirname, 'final.mdl'), os.path.join(destination, 'final.mdl'))
             copyfile(os.path.join(self.dirname, 'final.occs'), os.path.join(destination, 'final.occs'))
-            copyfile(os.path.join(self.dirname, 'tree'),  os.path.join(destination, 'tree'))
+            copyfile(os.path.join(self.dirname, 'tree'), os.path.join(destination, 'tree'))
 
     def export_triphone_fmllr_model(self, destination):
         """
         Add file into archive
         """
-        os.makedirs(destination, exist_ok = True)
+        os.makedirs(destination, exist_ok=True)
         copy(os.path.join(self.dirname, 'final.mdl'), destination)
         copy(os.path.join(self.dirname, 'final.occs'), destination)
         copy(os.path.join(self.dirname, 'tree'), destination)
